lambdas/airport-sync/requirements.txt
glide-py==1.0.0
httpx==0.27.0
